add roles guard

+++ src/user/user.controller.ts
@@ -1,36 +1,59 @@
 import {
-    ClassSerializerInterceptor,
     Controller,
     Delete,
     Get,
     Param,
     ParseUUIDPipe,
-    UseInterceptors,
+    UseGuards,
+    ForbiddenException,
+    NotFoundException,
 } from '@nestjs/common';
 import { UserService } from '@user/user.service';
 import { UserResponse } from '@user/responses';
-import { CurrentUser } from '@common/decorators';
+import { CurrentUser, Roles } from '@common/decorators';
 import { JwtPayload } from '@auth/interfaces';
+import { RolesGuard } from '@auth/guards/role.guard';
+import { JwtAuthGuard } from '@auth/guards/jwt-auth.guard';
+import { Role } from '@prisma/client';
 
 @Controller('user')
+@UseGuards(JwtAuthGuard)
 export class UserController {
     constructor(private readonly userService: UserService) {}
 
+    @UseGuards(RolesGuard)
+    @Roles(Role.ADMIN)
     @Get()
     async findAllUsers() {
-        return this.userService.findAll();
+        const users = await this.userService.findAll();
+        return users.map((user) => new UserResponse(user));
     }
 
-    @UseInterceptors(ClassSerializerInterceptor)
     @Get(':idOrEmail')
     async findOneUser(@Param('idOrEmail') idOrEmail: string) {
         const user = await this.userService.finOne(idOrEmail);
-
+        if (!user) {
+            throw new NotFoundException('User not found');
+        }
         return new UserResponse(user);
     }
 
     @Delete(':id')
     async deleteUser(@Param('id', ParseUUIDPipe) id: string, @CurrentUser() user: JwtPayload) {
-        return this.userService.delete(id, user);
+        if (user.id !== id && !user.roles.includes(Role.ADMIN)) {
+            throw new ForbiddenException('You are not allowed to delete this user');
+        }
+        const deletedUser = await this.userService.delete(id, user);
+        if (!deletedUser) {
+            throw new NotFoundException('User not found');
+        }
+        return { message: 'User deleted successfully' };
+    }
+
+    @UseGuards(RolesGuard)
+    @Roles(Role.ADMIN)
+    @Get('me')
+    me(@CurrentUser() user: JwtPayload) {
+        return user;
     }
 }
